requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr!=2.1.0,>=2.0.0 # Apache-2.0
openstacksdk>=0.29.0 # Apache-2.0
requests>=2.18.4 # Apache-2.0
PrettyTable<0.8,>=0.7.2 # BSD
